*7.2* huge opt. for RawToInt: added asm and PurePascal version for RawByteStrings of John O'Harrow fixed FPC Compiling Issue
git-svn-id: http://svn.code.sf.net/p/zeoslib/code-0/branches/testing-7.2@2440 d681b350-854c-439c-a81d-c0ce20710203

+++ src/dbc/ZDbcPostgreSqlStatement.pas
@@ -1198,7 +1198,7 @@ begin
   FPlainDriver := PlainDriver;
   ResultSetType := rtScrollInsensitive;
   FConnectionHandle := Connection.GetConnectionHandle;
-  FPlanName := IntToString(Hash(ASQL)+Cardinal(FStatementId)+NativeUInt(FConnectionHandle));
+  FPlanName := IntToString(Int64(Hash(ASQL)+Cardinal(FStatementId)+NativeUInt(FConnectionHandle)));
 end;
 
 procedure TZPostgreSQLCAPIPreparedStatement.Prepare;
